requirements.txt
kafka-python==2.0.1
requests==2.22.0
